fix(codepipeline): cross-env pipeline cannot be created in `Stage` (#17730)
Because a cross-environment pipeline cannot be created in `Stage`,
it cannot be deployed using CDK Pipelines.

The error is:

```
Error: You cannot add a dependency from 'AAA' (in Stage 'BBB') to 'CCC' (in the App): dependency cannot cross stage boundaries
```

Root cause is that the `Pipeline` construct creates a support stack
in the `App` scope, which is outside its containing `Stage`, and hence
the dependency crosses stage boundaries.

Fixes #17643.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*