feat(ecs-service-extensions): Auto scaling for Queue Extension (#17430)
----
This PR adds target tracking auto scaling policy for the the SQS Queues provided to and created by the `QueueExtension` (in the `useService()` hook).

The auto scaling is based on `backlogPerTask` custom metric which is emitted by an AWS Lambda Function. The PR also contains this Lambda Function and its tests.

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*