chore: avoid runtime dependency between cfnspec and pkglint
Build is currently failing in the pipeline because of this.

Since `pkglint` is private, `cfnspec` (which is public) cannot take a runtime dependency to it. But we still want to generate the same README format for CFN modules, so instead we just statically copy the relevant files from pkglint during build. Ugly but works.