chore: avoid runtime dependency between cfnspec and pkglint (#17751)
Our build pipeline is currently failing because `cfnspec` (which is a public package) takes a runtime dependency on `pkglint` (which is a private package). This was introduced in https://github.com/aws/aws-cdk/pull/17696.

To resolve this, we moved the functionality that generates new L1 library code from `lib/` to `build-tools/` since it's only needed in CDK build context (technically this is breaking public API, but we could not see any use case for external users to generate L1 modules in CDK-repository format).

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*