chore(apigatewayv2-authorizers): re-organize authorizer api (#17772)
This is a follow on to a previous commit 29039e8b.

Update the ergonomics of the authorizer construct APIs to be aligned
with the rest of the module, specifically the integration construct
APIs.

The API now takes the construct id and the integration target as part of
the constructor, instead of in the props class.

In most cases, except in the case of jwt, all properties in the props
struct become optional, which improves API ergonomics.
It also removes the need for `authorizerName` property to be required.

BREAKING CHANGE: The default value for the prop `authorizerName`
in `HttpJwtAuthorizerProps` has changed.
* **apigatewayv2-authorizers:** `HttpJwtAuthorizer` now takes the
  construct id and the target jwt issuer as part of its constructor.
* **apigatewayv2-authorizers:** `HttpLambdaAuthorizer` now takes
  the construct id and the target lambda function handler as part of
  its constructor.
* **apigatewayv2-authorizers:** The default value for the prop
  `authorizerName` in `HttpUserPoolAuthorizerProps` has changed.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*