chore: re-allow monocdk deep imports
#17707 added an `exports` section to the `package.json` files of `aws-cdk-lib` and `monocdk` in order to prevent users from accidentally "deep importing" modules.

This broke a large number of `monocdk` users. Since these teams are going to migrate to v2 (`aws-cdk-lib`) soon anyway, we decided to revert this change for `monocdk` so they are unblocked in the meantime.

To do this, we introduced two new configuration options for `ubergen`: `libRoot`, which can be used to control where submodules are going to be collected (defaults to the root of the uber package) and `explicitExports` which controls whether the `exports` section is added or not (defaults to true).