chore: fix check-api-compatibility after change in distribution tags
The change in #17823 introduced the `latest-1` distribution tags for the shared
packages published in v1. This change broke the `check-api-compat` script
because -- for bump builds -- it compares compatibility against the tag in
npm. With a new tag introduced, this doesn't work. Instead, simply default to
the latest at the same major version; this is equivalent and more robust.