fix(stepfunctions): prefixes not appended to states in parallel branches (#17806)
There is an order mismatch issue when calling `Parallel.branch()`. `branch` used to immediately create the state graph of the branch which locks in the `stateIds` of the branch states. This fails to take into account the following scenario: 

```ts
class ParallelMachineFragment extends stepfunctions.StateMachineFragment {
  public readonly startState: stepfunctions.State;
  public readonly endStates: stepfunctions.INextable[];

  constructor(scope: Construct, id: string) {
    super(scope, id);

    const step1 = new stepfunctions.Pass(this, 'Step 1');
    const parallelState = new stepfunctions.Parallel(this, 'Parallel State');
    const chain = parallelState.branch(step1);
    this.startState = parallelState;
    this.endStates = [chain];
  }
}

class MyStack extends Stack {
  constructor(scope: Construct, id: string) {
    const fragment1 = new ParallelMachineFragment(this, 'Fragment 1').prefixStates();
    const fragment2 = new ParallelMachineFragment(this, 'Fragment 2').prefixStates();
    
    new stepfunctions.StateMachine(stack, 'State Machine', {
      definition: fragment1.next(fragment2),
    });
  }
}
```

Here, the branches in the `parallelState` do not get prefixes before they are evaluated, causing an error since they have the same `stateId`.

The fix is to make the state graph creation late-bound; it now happens when we call `Parallel.bindToGraph()`. At this point, it will know of any prefixes. 

Fixes #17354. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*