fix(ecs-patterns): removeDefaultDesiredCount feature flag not expired properly (#17865)
The `@aws-cdk/aws-ecs-patterns:removeDefaultDesiredCount` feature flag was
expired as part of V2; however, the way the feature flag was accessed meant that
the flag wasn't handled the same way all other expired flags were, leading to
the incorrect behavior.

This fixes the access and evaluation of the flag to match intended behavior, and
updates the tests appropriately.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*