chore: add integration test to test a construct with a builtin lambda (#17571)
This adds a new integration test that deploys an s3.Bucket with
autoDeleteObjects set to true. The autoDeleteObjects feature deploys a
Nodejs Lambda backed Custom Resource.

Lambda backed custom resources that are included as part of CDK
constructs are compiled and bundled as part of the construct library.
There are scenarios where this compiled source code (e.g.
__entrypoint__.js) could be modified by the build process and cause the
lambda execution to fail.

This integration test should catch those instances. If the lambda
function throws errors during execution the CustomResource will
eventually fail. In the integration test this will result in a test
timeout and failure.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*