feat(codecommit): allow initializing a Repository with contents (#17968)
This repo introduces new properties to the constructor signature of the codecommit L2 Repository.
It allows users to upload code when creating a codecommit repository by leveraging the aws-s3-assets lib.
The user is able to upload whole directories at the moment.

The behaviour has a unit tests, and an integration test, which is passing (verified manually as well).

Closes #17967, provides a possible fix to #16958

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*