fix(apigateway): race condition between Stage and CfnAccount (#18011)
When `cloudWatchRole` is enabled, a `CfnAccount` is created for it. Since there is no explicit dependency between the the stages and the account, CloudFormation may deploy them in the wrong order, causing the deployment to fail.

Add an explicit dependency between `Stage`s (whether defined by the user or created automatically) and the CloudWatch `CfnAccount`, if it exists.

Fixes https://github.com/aws/aws-cdk/issues/10722.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*