chore: fix `yarn.lock` file
There are some missing entries in the `yarn.lock` file, breaking the
validation that we have on this.

The validation script should have always failed the build (even on the
build server), but it's not because it didn't properly handle uncaught
rejected promises, and whether that would fail or not depended on the
Node.js version.

Rectify both issues.