chore: fix `yarn.lock` file (#18068)
There are some missing entries in the `yarn.lock` file, breaking the
validation that we have on this.

The validation script should have always failed the build (even on the
build server), but it was not because it didn't properly handle uncaught
rejected promises, and whether that would fail or not depended on the
Node.js version.

Rectify both issues.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*