chore(aws-cdk-lib): minify aws-cdk-lib sources
This PR takes one step toward improving the load times for
`aws-cdk-lib`. Post-build, `esbuild` is used to minify the source and move
source maps to external files. In local testing, this changed the average time
for loading `aws-cdk-lib` from ~1110ms to ~830ms (25% reduction), and the size
of the locally-packed JS-only source from 53MB to 45MB.

related #18036