chore(v2): literate examples missing from v2
The literate examples are `test/*.lit.ts` files that are included
in the `README`s by jsii.

Ubergen didn't copy `test` files, because there were issues with the
test dependency `@aws-cdk/assert` which couldn't be included into
`aws-cdk-lib`.

In this PR, add copying of only the literate source files to `ubergen`
so that we can include the examples again in READMEs. It is safe since
those are integration tests that don't depend on `@aws-cdk/assert`
(which is only used by unit tests).