chore: don't bump empty releases
v2.3.0 was an empty release compared to v2.2.0, because the forward
merge had been failing.

It's pointless to release an empty release, so have the bump script
check for this situation, and not do anything.

Slightly complicated from the obvious logic which would only check
commit history: in practice the merge-back commit gets added on
top. So we check the actual file changes since the previous release.