fix(events): event bus name only generated if no props passed
Since the `EventBus` resource requires a name, CDK will generate
one for you.

However, because of the way the logic was written, it would only
generate one for you if you did NOT pass a props object. In effect,
these two statements would have a different effect:

```ts
new EventBus(this, 'Bus');

new EventBus(this, 'Bus', {});
```

Fix that issue.

Fixes #18070.