chore: rename label for auto approvals (#18183)
The current label (`pr/auto-approve`) doesn't match the default label that `projen` uses (`auto-approve`). 

Our repository monitoring mechanisms use `pr/auto-approve` to detect stale PR's, while this is OK for the `jsii` and `aws-cdk` repositories, its not for basically all other repo's, since they are managed by projen. 

So we need to change the label our monitoring uses to `auto-approve`, this PR is a preparation for that, so we don't loose monitoring of this repo.  

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*