chore: update run-rosetta.sh (#18127)
This point of this PR is to streamline the docs build in `run-rosetta.sh`. We now
utilize the npm package [`cdk-generate-synthetic-examples`](https://github.com/cdklabs/cdk-generate-synthetic-examples) owned by yours truly.
We also run `rosetta extract` twice, to properly translate generated examples. We purposefully ignore infused examples for this step.

In addition, we remove the `generate-examples` project

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*