chore(v2): literate examples missing from v2 (#18118)
The literate examples are `test/*.lit.ts` files that are included
in the `README`s by jsii.

Ubergen didn't copy `test` files, because there were issues with the
test dependency `@aws-cdk/assert` which couldn't be included into
`aws-cdk-lib`.

In this PR, add copying of only the literate source files to `ubergen`
so that we can include the examples again in READMEs. It is safe since
those are integration tests that don't depend on `@aws-cdk/assert`
(which is only used by unit tests).

Updated some EKS integ tests to no longer depend on test files that
don't get copied by this mechanism (`./util.ts`).


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*