feat(cli): support for hotswapping Lambda Versions and Aliases (#18145)
This extends the existing hotswapping support for Lambda Functions to also work for Versions and Aliases.

Implementation-wise, this required quite a bit of changes, as Versions are immutable in CloudFormation,
and the only way you can change them is by replacing them;
so, we had to add the notion of replacement changes to our hotswapping logic
(as, up to this point, they were simply a pair of "delete" and "add" changes,
which would result in hotswapping determining it can't proceed,
and falling back to a full CloudFormation deployment).

I also modified the main hotswapping algorithm:
now, a resource change is considered non-hotswappable if all detectors return `REQUIRES_FULL_DEPLOYMENT` for it;
if at least one detector returns `IRRELEVANT`, we ignore this change.
This allows us to get rid of the awkward `EmptyHotswapOperation` that we had to use before in these situations.

I also made a few small tweaks to the printing messages added in #18058:
I no longer prefix them with the name of the service,
as now hotswapping can affect different resource types, and it looked a bit awkward with that prefix present.

Fixes #17043

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*