chore(cfnspec): consume CloudFormation specification in parts
Whenever there are errors in the CloudFormation specification, we
currently have to fail the build and can't consume anything.

To ensure we make some progress, apply the following strategy instead:

- Split the spec into fragments, on a per-service basis.
- Consume those per-service spec updates that are valid; if updates
  are invalid, we will leave them at the old version.

This will produce an always-building spec, of which certain parts
may be outdated. Report the outdated parts in the CHANGELOG.

Notifying the CloudFormation team about spec errors is an out-of-band
process, and out of scope of this PR.

As a side effect of this work, formalize the spec manipulation we do
with JSON and patch files into a mini-standard called "JSON Patch
Stacks", and add some tools to operate on them.